server/requirements.txt
flask>=2.0.0
pyperclip>=1.8.0
pywin32>=300
python-dotenv>=0.19.0
